+++ index.js
@@ -1,304 +1,192 @@
 #!/usr/bin/env node
 
+const path = require("path");
+const fs = require("fs");
+
 const Jimp = require("jimp");
-const argv = require("yargs").options({
+const yargs = require("yargs");
+const svgToPng = require("svg-to-png");
+const _cliProgress = require('cli-progress');
+
+const { outputSizes, totalSizesLenght } = require("./output-sizes");
+
+
+const options = {
     "c": {
-        alias:   "color",
+        alias: "color",
         default: "448AFF",
-        type:    "string",
+        type: "string",
     },
     "i": {
-        alias:   "icon",
-        default: `${__dirname}/assets/default-icon.png`,
-        type:    "string",
+        alias: "icon",
+        default: "assets/default-icon.png",
+        type: "string",
     },
     "o": {
-        alias:   "output",
+        alias: "output",
         default: "media",
-        type:    "string",
+        type: "string",
     },
-}).argv;
+}
+const argv = yargs.options(options).argv;
 
-const bg_color  = Jimp.cssColorToHex(`#${argv.color}`);
-const icon_path = argv.icon;
+for (const i in options) {
+    if (options[i].default === argv[i]) {
+        console.info(`INFO: Using default value of '${options[i].alias}' (${argv[i]})`);
+    }
+}
 
-const output_sizes = {
-    "startup-image": {
-        "ios": [
-            {
-                dimensions:   [640, 960], // iPhone 4
-                scale_factor: 2,
-                mode:         "default",
-            },
-            {
-                dimensions:   [640, 1136], // iPhone 5
-                scale_factor: 2,
-                mode:         "default",
-            },
-            {
-                dimensions:   [750, 1334], // iPhone 6, 6S, 7, 8
-                scale_factor: 2,
-                mode:         "default",
-            },
-            {
-                dimensions:   [828, 1792], // iPhone XR
-                scale_factor: 2,
-                mode:         "default",
-            },
-            {
-                dimensions:   [1125, 2436], // iPhone X, XS
-                scale_factor: 2,
-                mode:         "default",
-            },
-            {
-                dimensions:   [1242, 2208], // iPhone 6+, 6S+, 7+, 8+
-                scale_factor: 2,
-                mode:         "default",
-            },
-            {
-                dimensions:   [1242, 2688], // iPhone XS Max
-                scale_factor: 2,
-                mode:         "default",
-            },
-            {
-                dimensions:   [768, 1024], // iPad 1, 2, Mini
-                scale_factor: 3,
-                mode:         "default",
-            },
-            {
-                dimensions:   [1536, 2048], // iPad 3, 4, Air, Mini 2, Pro 9.7
-                scale_factor: 3,
-                mode:         "default",
-            },
-            {
-                dimensions:   [1668, 2224], // iPad Pro 10.5
-                scale_factor: 3,
-                mode:         "default",
-            },
-            {
-                dimensions:   [1668, 2388], // iPad Pro 11
-                scale_factor: 3,
-                mode:         "default",
-            },
-            {
-                dimensions:   [2048, 2732], // iPad Pro 12.9
-                scale_factor: 3,
-                mode:         "default",
-            },
-        ],
-    },
-    "touch-icon": {
-        "ios": [
-            {
-                dimensions:   [76, 76],
-                scale_factor: 1.5,
-                mode:         "default",
-            },
-            {
-                dimensions:   [120, 120],
-                scale_factor: 1.5,
-                mode:         "default",
-            },
-            {
-                dimensions:   [152, 152],
-                scale_factor: 1.5,
-                mode:         "default",
-            },
-            {
-                dimensions:   [167, 167],
-                scale_factor: 1.5,
-                mode:         "default",
-            },
-            {
-                dimensions:   [180, 180],
-                scale_factor: 1.5,
-                mode:         "default",
-            },
-            {
-                dimensions:   [1024, 1024],
-                scale_factor: 1.5,
-                mode:         "default",
-            },
-        ],
-    },
-    "spotlight-icon": {
-        "ios": [
-            {
-                dimensions: [120, 120],
-                scale_factor: 1.5,
-                mode: "default",
-            },
-            {
-                dimensions: [80, 80],
-                scale_factor: 1.5,
-                mode: "default",
-            },
-            {
-                dimensions: [58, 58],
-                scale_factor: 1.5,
-                mode: "default",
-            },
-        ]
-    },
-    "settings-icon": {
-        "ios": [
-            {
-                dimensions: [87, 87],
-                scale_factor: 1.5,
-                mode: "default",
-            },
-            {
-                dimensions: [58, 58],
-                scale_factor: 1.5,
-                mode: "default",
-            },
-        ]
-    },
-    "splash-icon": {
-        "android": [
-            {
-                dimensions:   [512, 512],
-                scale_factor: 1,
-                mode:         "transparent",
-            },
-        ],
-    },
-    "launcher-icon": {
-        "android": [
-            {
-                dimensions:   [48, 48],
-                scale_factor: 1.75,
-                mode:         "circle",
-            },
-            {
-                dimensions:   [72, 72],
-                scale_factor: 1.75,
-                mode:         "circle",
-            },
-            {
-                dimensions:   [96, 96],
-                scale_factor: 1.75,
-                mode:         "circle",
-            },
-            {
-                dimensions:   [144, 144],
-                scale_factor: 1.75,
-                mode:         "circle",
-            },
-            {
-                dimensions:   [192, 192],
-                scale_factor: 1.75,
-                mode:         "circle",
-            },
-        ],
-    },
-    "notification-icon": {
-        "android": [
-            {
-                dimensions: [24, 24],
-                scale_factor: 1,
-                mode: "transparent",
-            },
-            {
-                dimensions: [36, 36],
-                scale_factor: 1,
-                mode: "transparent",
-            },
-            {
-                dimensions: [48, 48],
-                scale_factor: 1,
-                mode: "transparent",
-            },
-            {
-                dimensions: [72, 72],
-                scale_factor: 1,
-                mode: "transparent",
-            },
-            {
-                dimensions: [96, 96],
-                scale_factor: 1,
-                mode: "transparent",
-            },
-        ],
-        "ios": [
-            {
-                dimensions: [60, 60],
-                scale_factor: 1.5,
-                mode: "default",
-            },
-            {
-                dimensions: [40, 40],
-                scale_factor: 1.5,
-                mode: "default",
-            },
-        ],
-    },
-};
+const bgColor = Jimp.cssColorToHex(`#${argv.color}`);
 
-// Generate images
-Object.keys(output_sizes).forEach((name_prefix) => {
-    Object.keys(output_sizes[name_prefix]).forEach((platform) => {
-        output_sizes[name_prefix][platform].forEach((settings) => {
-            // set the default background in an array
-            const backgrounds = [new Jimp(settings.dimensions[0], settings.dimensions[1], bg_color)];
+new Promise(async (resolve) => {
+    // verify if icon are svg
+    if (path.extname(argv.icon) === '.svg') {
+        // get path of icon
+        const iconFolder = path.dirname(argv.icon);
+        const iconName = path.basename(argv.icon, '.svg');
 
-            // determine the icon size based on the x dimension divided by the scaling factor
-            const icon_size = Math.ceil(settings.dimensions[0] / settings.scale_factor);
+        // get fullpath of temp file
+        const tempIcon = `${iconName}.png`;
 
-            // if dimension aren't equal, add a rotated background to the array
-            if (settings.dimensions[0] !== settings.dimensions[1]) {
-                backgrounds.push(new Jimp(settings.dimensions[1], settings.dimensions[0], bg_color));
-            }
+        console.warn(`\nWARN: Please don't delete ${tempIcon}. We'll do it for you when we're done.\n`);
 
-            Jimp.read(icon_path).then((icon) => {
-                icon.resize(icon_size, Jimp.AUTO);
+        // create temp png of svg
+        await svgToPng.convert(path.resolve(argv.icon), iconFolder, {
+            defaultHeight: 8000,
+            defaultWidth: 8000,
+        });
 
-                backgrounds.forEach((background) => {
-                    // determine where to position the icon for it to appear centered
-                    const icon_position = [Math.ceil((background.bitmap.width / 2) - icon_size / 2), Math.ceil((background.bitmap.height / 2) - icon_size / 2)];
+        resolve({
+            iconPath: path.resolve(path.join(iconFolder, tempIcon)), 
+            bySVG: true
+        });
+    } else {
+        console.warn(`\nWARN: Please don't delete ${path.basename(argv.icon)}.\n`);
+        resolve({
+            iconPath: path.resolve(argv.icon),
+            bySVG: false
+        });
+    }
+}).then((params, bySVG) => {
+    // create a new progress bar instance and use shades_classic theme
+    const progress = new _cliProgress.Bar({
+        format: 'Progress [{bar}] {percentage}% ({value}/{total}) | ETA: {eta}s | Status: {status}'
+    });
+    
+    // start the progress bar with a total value of 'totalSizesLenght' and start value of 0
+    progress.start(totalSizesLenght, 0, {
+        status: 'Starting...'
+    });
+    let currentProgress = 0;
 
-                    // set the background to transparent if opaque is false
-                    if (settings.mode === "transparent") {
-                        background.opacity(0);
-                    }
 
-                    // check if the image needs to be masked
-                    new Promise((resolve) => {
-                        // adjust the background if mode is circle
-                        if (settings.mode === "circle") {
-                            // read the circle mask
-                            Jimp.read(`${__dirname}/assets/circle-mask.png`).then((mask) => {
-                                // reszie the circle mask to match the background
-                                mask.resize(background.bitmap.width, background.bitmap.height);
+    // Generate images
+    for (const namePrefix in outputSizes) {
+        const currentNamePrefix = outputSizes[namePrefix];
 
-                                // apply the circle mask the background
-                                background.mask(mask, 0, 0);
+        for (const platform in currentNamePrefix) {
+            const currentPlatform = currentNamePrefix[platform];
 
-                                // emboss
-                                background.convolute([[-2, -1, 0], [-1, 1, 1], [0, 1, 2]]);
+            currentPlatform.forEach((settings) => {
+                // set the default background in an array
+                const backgrounds = [new Jimp(settings.dimensions[0], settings.dimensions[1], bgColor)];
 
-                                // read the shadow
-                                Jimp.read(`${__dirname}/assets/circle-shadow.png`).then((shadow) => {
-                                    // resize the shadow to match the background
-                                    shadow.resize(background.bitmap.width, background.bitmap.height);
+                // determine the icon size based on the x dimension divided by the scaling factor
+                const icon_size = Math.ceil(settings.dimensions[0] / settings.scale_factor);
 
-                                    // layer the shadow below the background
-                                    shadow.composite(background, 0, 0);
+                // if dimension aren't equal, add a rotated background to the array
+                if (settings.dimensions[0] !== settings.dimensions[1]) {
+                    backgrounds.push(new Jimp(settings.dimensions[1], settings.dimensions[0], bgColor));
+                }
 
-                                    // return the background for final write
-                                    resolve(shadow);
-                                });
-                            });
-                        } else {
-                            // immediately resolve with the existing background if mode is not circle
-                            resolve(background);
+                const finalIconLocation = `${argv.output}/${platform}/`;
+                progress.update(currentProgress, {
+                    status: `${finalIconLocation}...`
+                });
+
+                Jimp.read(params.iconPath).then((icon) => {
+                    icon.cover(icon_size, icon_size);
+
+                    backgrounds.forEach((background) => {
+                        const finalIconSize = `${background.bitmap.width}x${background.bitmap.height}`;
+                        const finalIconFile = `${finalIconLocation}${namePrefix}-${finalIconSize}.png`;
+
+
+                        progress.update(currentProgress, {
+                            status: `${finalIconFile}...`
+                        });
+                        // determine where to position the icon for it to appear centered
+                        const icon_position = [Math.ceil((background.bitmap.width / 2) - icon_size / 2), Math.ceil((background.bitmap.height / 2) - icon_size / 2)];
+
+                        // set the background to transparent if opaque is false
+                        if (settings.mode === "transparent") {
+                            background.opacity(0);
                         }
-                    }).then((background) => {
-                        // write the image
-                        background.composite(icon, icon_position[0], icon_position[1]).write(`${argv.output}/${platform}/${name_prefix}-${background.bitmap.width}x${background.bitmap.height}.png`);
+
+                        // check if the image needs to be masked
+                        new Promise((resolve) => {
+                            // adjust the background if mode is circle
+                            if (settings.mode === "circle") {
+                                // read the circle mask
+                                Jimp.read(`${__dirname}/assets/circle-mask.png`).then((mask) => {
+                                    // resize the circle mask to match the background
+                                    mask.resize(background.bitmap.width, background.bitmap.height);
+
+                                    // apply the circle mask the background
+                                    background.mask(mask, 0, 0);
+
+                                    // emboss
+                                    background.convolute([[-2, -1, 0], [-1, 1, 1], [0, 1, 2]]);
+
+                                    // read the shadow
+                                    Jimp.read(`${__dirname}/assets/circle-shadow.png`).then((shadow) => {
+                                        // resize the shadow to match the background
+                                        shadow.resize(background.bitmap.width, background.bitmap.height);
+
+                                        // layer the shadow below the background
+                                        shadow.composite(background, 0, 0);
+
+                                        // return the background for final write
+                                        resolve(shadow);
+                                    });
+                                });
+                            } else {
+                                // immediately resolve with the existing background if mode is not circle
+                                resolve(background);
+                            }
+                        }).then((background) => {
+
+
+                            // write the image
+                            background.composite(icon, icon_position[0], icon_position[1]).write(finalIconFile);
+
+                            // update the current value
+                            currentProgress += 1;
+                            progress.update(currentProgress, {
+                                status: `${finalIconFile} OK!`
+                            });
+
+                            // verify if is the last item
+                            if (currentProgress === totalSizesLenght) {
+                                progress.update(currentProgress, {
+                                    status: `Finished!`
+                                });
+            
+                                // delete temp png of svg
+                                if (params.bySVG) {
+                                    fs.unlinkSync(params.iconPath);
+                                }
+
+                                // stop the progress bar
+                                progress.stop();
+                            }
+                        });
                     });
                 });
             });
-        });
-    });
+
+
+
+        }
+    }
 });