+++ CHANGELOG.md
@@ -3,6 +3,36 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.0.10] - 2019-02-09
+### Changed
+- New style of progress bar
+
+## [1.0.9] - 2019-02-09
+### Changed
+- Progress bar
+
+## [1.0.8] - 2019-02-09
+### Changed
+- (issue #3) Generate from an SVG
+
+## [1.0.7] - 2019-02-09
+### Changed
+- (issue #4) Ensure non-square images are properly centered
+
+## [1.0.6] - 2019-02-09
+### Changed
+- (issue #5) Show a warning when a default value is used
+- Resolve icon and output path and simplify default value
+
+## [1.0.5] - 2019-02-09
+### Changed
+- Use for..in instead of Object.keys
+- Reindent
+- Fix changelog v1.0.4 (## Changed)
+
+## [1.0.4] - 2019-02-09
+### Changed
+- Create file for output-sizes and total length
 
 ## [1.0.3] - 2019-01-08
 ### Changed